+++ src/document/builders.js
@@ -110,34 +110,6 @@ function fill(parts) {
 
   return { type: DOC_TYPE_FILL, parts };
 }
-fill.builder = fillBuilder;
-
-function fillBuilder() {
-  const parts = [];
-  let pendingDocs = [];
-  function flush() {
-    parts.push(pendingDocs);
-    pendingDocs = [];
-  }
-
-  const self = {
-    append(doc) {
-      pendingDocs.push(doc);
-      return self;
-    },
-    appendLine(doc) {
-      flush();
-      pendingDocs = [];
-      parts.push(doc);
-      return self;
-    },
-    build() {
-      flush();
-      return fill(parts);
-    },
-  };
-  return self;
-}
 
 /**
  * @param {Doc} breakContents
